tip: major rewrite
part two of the rewrite saga.

*fixed statusbar texture
*try to reduce cpu usage by calling things direct instead of copying
once or twice into a locale every time OnUnit is called
*reworked targetlineto fire on UNIT_CHANGED event instead of updating
every 0.1 seconds. This is slower but a lot more cpu friendly
*splitted tooltip modifications for NPC units and Player units, also
have a shared function which calls the previous mentioned.
*cleaned up talent display function; a lot of old code removed

As a side note targetline wasn't the culprict for the fps issues, the
OnSetUnit function did create GameTooltipStatusbar textures every time
it was called, after time the usage kept building up. This is fixed
aswell.