itemlevel: split into dedicated .lua's
This commit splits up the ilvl.lua into different individual .lua's so it is more readable. This also adresses some issues from the old implementation where i was not happy with an implementation either with this or that setup. So basically now we can decide how to adress certain usecases and this also paves the way for easier Bag Addon support implementations.
See #12

fixes #13
fixes #20